- Add ErrorHandler Struct - Add RealmParser Struct - Change Post model from struct to class - Separate RealmPostRepositoryImp to new swift file
