+++ Offline Clean Architecture/PostRepositoryImp.swift
@@ -8,18 +8,9 @@
 
 import Foundation
 import RxSwift
-import RealmSwift
-import Moya_ModelMapper
 import Moya
-import ReachabilitySwift
-
-
 
 struct NetworkPostRepositoryImp : PostRepositrory {
-    
-    
-    
-    
     
     func allPosts() -> Observable<[Post]> {
         
@@ -27,68 +18,17 @@ struct NetworkPostRepositoryImp : PostRepositrory {
         return p.request(.posts).map({ (response) -> [Post] in
             
             do {
-                let result = try response.filterSuccessfulStatusCodes()
-                let mappingResult = try result.mapArray() as [Post]
-                let realmObjects = Post.realmObjects(items: mappingResult)
-                
-                let realm = try! Realm()
-                try realm.write {
-                    realm.add(realmObjects, update: true)
-                }
-                return mappingResult
+                let mappingResult = try response.filterSuccessfulStatusCodes().mapArray() as [Post]
+                return try RealmParser().writeObject(realmObject: mappingResult)
                 
             }catch (let error) {
-                
-                if case MoyaError.underlying(let error) = error {
-                    print(error)
-                }else if case MoyaError.jsonMapping(let e) = error {
-                    print("ERROR : \(error), jsonMapping: \(e)")
-                }else if case MoyaError.statusCode(let status) = error {
-                    print("ERROR : \(error), Status: \(status)")
-                }else if case MoyaError.requestMapping(let value) = error {
-                    print("ERROR : \(error), requestMapping: \(value)")
-                }
-                throw error
+                throw ErrorHandler().handleError(error: error)
             }
         })
     }
     
     
-}
-
-
-
-struct RealmPostRepositoryImp : PostRepositrory {
     
-    func allPosts() -> Observable<[Post]> {
-        
-        return Observable.create({ (observer) -> Disposable in
-            
-            do {
-                
-                let realm = try Realm()
-                let items = realm.objects(PostRealmObject.self)
-                
-
-                let posts : [Post] = items.map({ (item) -> Post in
-                    let post = Post(userId: item.userId, id: item.id, title: item.title, body: item.body)
-                    return post
-                })
-                
-              
-                observer.onNext(posts)
-                observer.on(.completed)
-            }catch (let error) {
-                observer.onError(error)
-                observer.on(.completed)
-                
-                
-            }
-            return Disposables.create()
-        })
-        
-        
-        
-    }
     
 }
+
